Send battery level to server
Periodic workers don't have a failed state, so the notification listener
doesn't know if the job has failed or not. Because of that I used a
periodic worker for triggering a one time worker that sends the battery
level.
While charging the frequency is lowered to 10 - 15 minutes, if the
battery isn't low.

Closes #269

Signed-off-by: mueller-ma <mueller-ma@users.noreply.github.com>